+++ pyfakefs/tests/fake_pathlib_test.py
@@ -22,33 +22,24 @@ python docs.
 
 import errno
 import os
+import pathlib
 import stat
 import sys
 import unittest
 
-from pyfakefs.extra_packages import pathlib, pathlib2
 from pyfakefs.fake_filesystem import is_root
 
 from pyfakefs import fake_pathlib, fake_filesystem
+from pyfakefs.helpers import IS_PYPY
 from pyfakefs.tests.test_utils import RealFsTestCase
 
 is_windows = sys.platform == 'win32'
 
 
-def skip_if_pathlib_36_not_available():
-    if sys.version_info < (3, 6) and not pathlib2:
-        raise unittest.SkipTest('Changed behavior in Python 3.6')
-
-
-def skip_if_pathlib_36_is_available():
-    if sys.version_info >= (3, 6) or pathlib2:
-        raise unittest.SkipTest('Changed behavior in Python 3.6')
-
-
 class RealPathlibTestCase(RealFsTestCase):
     def __init__(self, methodName='runTest'):
         super(RealPathlibTestCase, self).__init__(methodName)
-        self.pathlib = pathlib or pathlib2
+        self.pathlib = pathlib
         self.path = None
 
     def setUp(self):
@@ -238,8 +229,8 @@ class FakePathlibPurePathTest(RealPathlibTestCase):
         with self.assertRaises(ValueError):
             self.path('passwd').relative_to('/usr')
 
-    @unittest.skipIf(sys.version_info < (3, 9) or pathlib2,
-                     'readlink new in Python 3.9')
+    @unittest.skipIf(sys.version_info < (3, 9),
+                     'is_relative_to new in Python 3.9')
     def test_is_relative_to(self):
         path = self.path('/etc/passwd')
         self.assertTrue(path.is_relative_to('/etc'))
@@ -388,8 +379,6 @@ class FakePathlibFileObjectPropertyTest(RealPathlibTestCase):
         # we get stat.S_IFLNK | 0o755 under MacOs
         self.assertEqual(link_stat.st_mode, stat.S_IFLNK | 0o777)
 
-    @unittest.skipIf(sys.platform == 'darwin',
-                     'Different behavior under MacOs')
     def test_lchmod(self):
         self.skip_if_symlink_not_supported()
         file_stat = self.os.stat(self.file_path)
@@ -400,8 +389,26 @@ class FakePathlibFileObjectPropertyTest(RealPathlibTestCase):
         else:
             self.path(self.file_link_path).lchmod(0o444)
             self.assertEqual(file_stat.st_mode, stat.S_IFREG | 0o666)
-            # we get stat.S_IFLNK | 0o755 under MacOs
-            self.assertEqual(link_stat.st_mode, stat.S_IFLNK | 0o444)
+            # the exact mode depends on OS and Python version
+            self.assertEqual(link_stat.st_mode & 0o777700,
+                             stat.S_IFLNK | 0o700)
+
+    @unittest.skipIf(sys.version_info < (3, 10),
+                     "follow_symlinks argument new in Python 3.10")
+    def test_chmod_no_followsymlinks(self):
+        self.skip_if_symlink_not_supported()
+        file_stat = self.os.stat(self.file_path)
+        link_stat = self.os.lstat(self.file_link_path)
+        if os.chmod not in os.supports_follow_symlinks or IS_PYPY:
+            with self.assertRaises(NotImplementedError):
+                self.path(self.file_link_path).chmod(0o444,
+                                                     follow_symlinks=False)
+        else:
+            self.path(self.file_link_path).chmod(0o444, follow_symlinks=False)
+            self.assertEqual(file_stat.st_mode, stat.S_IFREG | 0o666)
+            # the exact mode depends on OS and Python version
+            self.assertEqual(link_stat.st_mode & 0o777700,
+                             stat.S_IFLNK | 0o700)
 
     def test_resolve(self):
         self.create_dir(self.make_path('antoine', 'docs'))
@@ -417,11 +424,6 @@ class FakePathlibFileObjectPropertyTest(RealPathlibTestCase):
             self.path(
                 self.os.path.realpath(
                     self.make_path('antoine', 'setup.py'))))
-
-    def test_resolve_nonexisting_file(self):
-        skip_if_pathlib_36_is_available()
-        path = self.path('/foo/bar')
-        self.assert_raises_os_error(errno.ENOENT, path.resolve)
 
     def test_stat_file_in_unreadable_dir(self):
         self.check_posix_only()
@@ -448,25 +450,7 @@ class FakePathlibFileObjectPropertyTest(RealPathlibTestCase):
             path = str(list(iter)[0])
             self.assertTrue(path.endswith('some_file'))
 
-    @unittest.skipIf(not is_windows, 'Windows specific behavior')
-    def test_resolve_file_as_parent_windows(self):
-        skip_if_pathlib_36_is_available()
-        self.check_windows_only()
-        self.create_file(self.make_path('a_file'))
-        path = self.path(self.make_path('a_file', 'this can not exist'))
-        self.assert_raises_os_error(errno.ENOENT, path.resolve)
-
-    @unittest.skipIf(is_windows, 'POSIX specific behavior')
-    def test_resolve_file_as_parent_posix(self):
-        skip_if_pathlib_36_is_available()
-        self.check_posix_only()
-        self.create_file(self.make_path('a_file'))
-        path = self.path(
-            self.make_path('', 'a_file', 'this can not exist'))
-        self.assert_raises_os_error(errno.ENOTDIR, path.resolve)
-
-    def test_resolve_nonexisting_file_after_36(self):
-        skip_if_pathlib_36_not_available()
+    def test_resolve_nonexisting_file(self):
         path = self.path(
             self.make_path('/path', 'to', 'file', 'this can not exist'))
         self.assertEqual(path, path.resolve())
@@ -557,6 +541,19 @@ class FakePathlibPathFileOperationTest(RealPathlibTestCase):
         self.assertTrue(self.os.path.exists(path_name))
         self.check_contents(path_name, 'ανοησίες'.encode('greek'))
 
+    @unittest.skipIf(sys.version_info < (3, 10),
+                     "newline argument new in Python 3.10")
+    def test_write_with_newline_arg(self):
+        path = self.path(self.make_path('some_file'))
+        path.write_text('1\r\n2\n3\r4', newline='')
+        self.check_contents(path, b'1\r\n2\n3\r4')
+        path.write_text('1\r\n2\n3\r4', newline='\n')
+        self.check_contents(path, b'1\r\n2\n3\r4')
+        path.write_text('1\r\n2\n3\r4', newline='\r\n')
+        self.check_contents(path, b'1\r\r\n2\r\n3\r4')
+        path.write_text('1\r\n2\n3\r4', newline='\r')
+        self.check_contents(path, b'1\r\r2\r3\r4')
+
     def test_read_bytes(self):
         path_name = self.make_path('binary_file')
         self.create_file(path_name, contents=b'Binary file contents')
@@ -644,7 +641,7 @@ class FakePathlibPathFileOperationTest(RealPathlibTestCase):
         self.assertTrue(self.os.path.exists(link_name))
         self.assertTrue(path.is_symlink())
 
-    @unittest.skipIf(sys.version_info < (3, 8) or pathlib2,
+    @unittest.skipIf(sys.version_info < (3, 8),
                      'link_to new in Python 3.8')
     def test_link_to(self):
         self.skip_if_symlink_not_supported()
@@ -658,20 +655,21 @@ class FakePathlibPathFileOperationTest(RealPathlibTestCase):
         self.assertFalse(path.is_symlink())
         self.assertEqual(2, self.os.stat(file_name).st_nlink)
 
-    @unittest.skipIf(sys.version_info < (3, 8) or not pathlib2,
-                     'pathlib2 has no link_to')
-    def test_pathlib2_does_not_have_link_to(self):
+    @unittest.skipIf(sys.version_info < (3, 10),
+                     'hardlink_to new in Python 3.10')
+    def test_hardlink_to(self):
         self.skip_if_symlink_not_supported()
         file_name = self.make_path('foo', 'bar.txt')
         self.create_file(file_name)
         self.assertEqual(1, self.os.stat(file_name).st_nlink)
-        link_name = self.make_path('link_to_bar')
+        link_path = self.path(self.make_path('link_to_bar'))
         path = self.path(file_name)
-        with self.assertRaises(AttributeError):
-            path.link_to(link_name)
-        self.assertFalse(self.os.path.exists(link_name))
+        link_path.hardlink_to(path)
+        self.assertTrue(self.os.path.exists(link_path))
+        self.assertFalse(path.is_symlink())
+        self.assertEqual(2, self.os.stat(file_name).st_nlink)
 
-    @unittest.skipIf(sys.version_info < (3, 9) or pathlib2,
+    @unittest.skipIf(sys.version_info < (3, 9),
                      'readlink new in Python 3.9')
     def test_readlink(self):
         self.skip_if_symlink_not_supported()
@@ -1014,7 +1012,22 @@ class FakePathlibUsageInOsFunctionsTest(RealPathlibTestCase):
     def test_stat(self):
         path = self.make_path('foo', 'bar', 'baz')
         self.create_file(path, contents='1234567')
-        self.assertEqual(self.os.stat(path), self.os.stat(self.path(path)))
+        self.assertEqual(self.os.stat(path), self.path(path).stat())
+
+    @unittest.skipIf(sys.version_info < (3, 10), "New in Python 3.10")
+    def test_stat_follow_symlinks(self):
+        self.check_posix_only()
+        directory = self.make_path('foo')
+        base_name = 'bar'
+        file_path = self.path(self.os.path.join(directory, base_name))
+        link_path = self.path(self.os.path.join(directory, 'link'))
+        contents = "contents"
+        self.create_file(file_path, contents=contents)
+        self.create_symlink(link_path, base_name)
+        self.assertEqual(len(contents),
+                         link_path.stat(follow_symlinks=True)[stat.ST_SIZE])
+        self.assertEqual(len(base_name),
+                         link_path.stat(follow_symlinks=False)[stat.ST_SIZE])
 
     def test_utime(self):
         path = self.make_path('some_file')